v1.0 May 2023 Extensive refactoring

+++ js/showlegs.js
@@ -34,7 +34,7 @@ async function initShowLegs() {
         // ** step 1:
         // ***********
         let startAt = getURIParameterByName("start");
-        let initialLegID = getURIParameterByName("id");
+        let startWithLegID = getURIParameterByName("id");
         if (startAt) {
             startAt = startAt.split(",");
             APP.map = MAP.init("ShowLegsMap", [startAt[0], startAt[1]], startAt[2], selectClickedLeg); //loc from params
@@ -57,8 +57,8 @@ async function initShowLegs() {
         /*-- ********** -*/
         // ** step 2:
         // ***********
-        if (initialLegID) {
-            let theWhereStr = 'id=eq.' + initialLegID;
+        if (startWithLegID) {
+            let theWhereStr = 'id=eq.' + startWithLegID;
             doLegsSearchAndShow(theWhereStr, true, true, true, null);
         }
         showSearchLegsForm(false, true, true); // in Utils.js
@@ -405,17 +405,40 @@ function editLeg(theLegs, ID) {
 
 /*-- Delete selected Leg  -*/
 async function deleteLeg(theLegs, ID) {
-    const theStr = `Are you REALLY sure you want to delete this Leg [id=${ID}]?\nThis action can NOT be undone!`;
-    if (confirm(theStr)) {
-        if (await DB.deleteLeg(ID)) { //removed from DB
-            //reload from DB:
-            await theLegs.loadFromDB(theLegs.getWhereStr(), APP.legstops); // re-use existing WHERE statement to reload same legs
-            UI.SetMessage("Deleted Leg. Now " + theLegs.getNumLegs() + " Legs found in DB. Click in map or table to alter selection...", workflowMsg);
-            theLegs.mapLegs(APP.legstops, false);
-            displayLegsInTable(theLegs, false);
-        } else {
-            UI.SetMessage("Deletion of Leg failed. Click in map or table to alter selection...", workflowMsg);
+    // Check if leg used in Journey: if so, cannot delete!
+    let canDelete = true;
+    let postUrl = '/journeys?select=id,legsarray';
+    // console.log(postUrl);
+    let resultJSON = await DB.query("GET", postUrl);
+    if (resultJSON.error === true) {
+        DB.giveErrorMsg(resultJSON);
+        canDelete = false
+    } else {
+        for (let journeyJSON of resultJSON.data) {
+            if (journeyJSON.legsarray.includes(ID)) {
+                canDelete = false;
+                const theStr = "Cannot delete Leg " + ID + ", it is part of Journey " + journeyJSON.id + "\n[click OK to see/edit this Journey, cancel to skip]"
+               if (confirm(theStr)) {
+                   window.open("./showjourneys.html?id=" + journeyJSON.id);
+               }
+            }
         }
+    }
+    if (canDelete) {
+        const theStr = `Are you REALLY sure you want to delete this Leg [id=${ID}]?\nThis action can NOT be undone!`;
+        if (confirm(theStr)) {
+            if (await DB.deleteLeg(ID)) { //removed from DB
+                //reload from DB:
+                await theLegs.loadFromDB(theLegs.getWhereStr(), APP.legstops); // re-use existing WHERE statement to reload same legs
+                UI.SetMessage("Deleted Leg. Now " + theLegs.getNumLegs() + " Legs found in DB. Click in map or table to alter selection...", workflowMsg);
+                theLegs.mapLegs(APP.legstops, false);
+                displayLegsInTable(theLegs, false);
+            } else {
+                UI.SetMessage("Deletion of Leg failed. Click in map or table to alter selection...", workflowMsg);
+            }
+        }
+    } else {
+        UI.SetMessage("Deletion of Leg cancelled. Click in map or table to alter selection...", workflowMsg);
     }
 }
